Add replacement support for browser telemetry
Uses the same mechanism as implemented for node.

This change also moves the shared logic to `utils.ts` and the
shared interface to `baseTelemetryReporter.ts`.

It updates the typings file appropriately.
