+++ src/common/util.ts
@@ -3,6 +3,7 @@
  *--------------------------------------------------------*/
 
 import * as vscode from "vscode";
+import { ReplacementOption } from "./baseTelemetryReporter";
 
 export const enum TelemetryLevel {
 	ON = "on",
@@ -33,5 +34,19 @@ export function getTelemetryLevel(): TelemetryLevel {
 		const config = vscode.workspace.getConfiguration(TELEMETRY_CONFIG_ID);
 		const enabled = config.get<boolean>(TELEMETRY_CONFIG_ENABLED_ID);
 		return enabled ? TelemetryLevel.ON : TelemetryLevel.OFF;
+	}
+}
+
+export function applyReplacements(data: Record<string, any>, replacementOptions: ReplacementOption[]) {
+	for (const key of Object.keys(data)) {
+		for (const option of replacementOptions) {
+			if (option.lookup.test(key)) {
+				if (option.replacementString !== undefined) {
+					data[key] = option.replacementString;
+				} else {
+					delete data[key];
+				}
+			}
+		}
 	}
 }
